03-mlops/terraform/requirements.txt
google-cloud-aiplatform
kfp
google-cloud-pipeline-components
